Tratando erro 404 utilizando o método findById

+++ src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -2,6 +2,7 @@ package br.com.alura.forum.controller;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 import javax.validation.Valid;
@@ -36,6 +37,7 @@ public class TopicosController {
 	
 	@Autowired
 	CursoRepository cursoRepository;
+	
 
     @GetMapping
     public List<TopicoDto> lista(String nomeCurso) {
@@ -51,6 +53,7 @@ public class TopicosController {
     	
     }
     
+    
     @PostMapping
     @Transactional
     public ResponseEntity<TopicoDto> cadastrar(@RequestBody @Valid TopicoForm form, UriComponentsBuilder uriBuilder) {
@@ -62,26 +65,47 @@ public class TopicosController {
     	  	
     }
     
+    
     @GetMapping("/{id}")
-    public DetalhesDoTopicoDto detalhar(@PathVariable Long id) {
-    	Topico topico = topicoRepository.getReferenceById(id);
-    	return new DetalhesDoTopicoDto(topico);
+    public ResponseEntity<DetalhesDoTopicoDto> detalhar(@PathVariable Long id) {
+    	//Topico topico = topicoRepository.getReferenceById(id);
+        //substituido pela linha abaixo c/ objeto de devolver Status: 404 Not Found se o id não existir.
+    	Optional<Topico> topico = topicoRepository.findById(id);
+    	if(topico.isPresent()) {
+    		return  ResponseEntity.ok(new DetalhesDoTopicoDto(topico.get()));
+    	}
     	
+    	return ResponseEntity.notFound().build();
     }
+    
     
     @PutMapping("/{id}")
     @Transactional
     public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form){
-    	Topico topico = form.atualizar(id, topicoRepository);
     	
-    	return ResponseEntity.ok(new TopicoDto(topico));    	
+    	Optional<Topico> optional = topicoRepository.findById(id);
+    	
+    	if(optional.isPresent()) {
+    		Topico topico = form.atualizar(id, topicoRepository);
+    		return ResponseEntity.ok(new TopicoDto(topico)); 
+    	}
+    	
+    	return ResponseEntity.notFound().build();
     }
-    
+    	
+        
     @DeleteMapping("/{id}")
     @Transactional
-    public ResponseEntity<?> remover(@PathVariable Long id) {    	
-    	topicoRepository.deleteById(id);    	    	
-    	return ResponseEntity.ok().build();
+    public ResponseEntity<?> remover(@PathVariable Long id) {  
+    	
+    	Optional<Topico> optional = topicoRepository.findById(id);
+    	
+    	if(optional.isPresent()) {
+    		topicoRepository.deleteById(id);    	    	
+        	return ResponseEntity.ok().build();
+    	}
+    	
+    	return ResponseEntity.notFound().build();
     }
 
 }
